+++ CHANGELOG.md
@@ -4,7 +4,9 @@ List of changes for this repo, including `atomic-cli`, `atomic-server` and `atom
 
 ## v0.21.0
 
-- Add JSON-AD serialization #79
+- Add JSON-AD serialization #79, use it in Commits
+- Servers are aware of their own URL #51
+- Improved CLI edit feature, more flexible (create new resources if none exist, fix newlines)
 
 ## v0.20.4